% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4.enrichment.R
\name{enrichment_individual}
\alias{enrichment_individual}
\alias{ORA_GO}
\alias{GSEA_GO}
\alias{ORA_KEGG}
\alias{GSEA_KEGG}
\alias{ORA_MKEGG}
\alias{GSEA_MKEGG}
\alias{ORA_Reactome}
\alias{GSEA_Reactome}
\title{Enrichment wrappers (individual)}
\usage{
ORA_GO(gene, simplify = T)

GSEA_GO(geneList, simplify = T)

ORA_KEGG(gene, input = "SYMBOL", ...)

GSEA_KEGG(geneList, input = "SYMBOL")

ORA_MKEGG(gene, input = "SYMBOL", ...)

GSEA_MKEGG(geneList, input = "SYMBOL")

ORA_Reactome(gene, input = "SYMBOL")

GSEA_Reactome(geneList, input = "SYMBOL")
}
\arguments{
\item{gene}{SYMBOL or ENTREZID vector depending on input}

\item{simplify}{Use clusterProfiler::simplify for GO terms}

\item{geneList}{Named numeric vector (SYMBOL or ENTREZID per input)}

\item{input}{"SYMBOL" or "ENTREZID"}

\item{organism}{KEGG organism code, default "hsa"}
}
\value{
enrichResult

gseaResult

enrichResult

gseaResult

enrichResult

gseaResult

enrichResult

gseaResult
}
\description{
ORA/GSEA wrappers for GO, KEGG, KEGG Module, and Reactome.
}
\details{
Enrichment wrappers (individual)

'''{R}
}
\section{Functions}{

\itemize{
\item \code{ORA_GO}, \code{GSEA_GO}
\item \code{ORA_KEGG}, \code{GSEA_KEGG}
\item \code{ORA_MKEGG}, \code{GSEA_MKEGG}
\item \code{ORA_Reactome}, \code{GSEA_Reactome}
}
}

\section{KEGG_vis_notes: https://yulab-smu.top/biomedical-knowledge-mining-book/clusterprofiler-kegg.html}{
library("pathview")
kk <- gseKEGG(geneList     = geneList,
keyType      = 'SYMBOL',
organism     = 'hsa',
minGSSize    = 120,
pvalueCutoff = 0.05,
verbose      = FALSE)
browseKEGG(kk, 'hsa04110')
}

\section{or}{
hsa04110 <- pathview(gene.data  = geneList,
pathway.id = "hsa04110",
species    = "hsa",
limit      = list(gene=max(abs(geneList)), cpd=1))
'''
}

\seealso{
leo_enrich
}
