% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4.enrichment.R
\name{leo_enrich}
\alias{leo_enrich}
\title{Enrichment analysis (Integrated)}
\usage{
leo_enrich(
  gene,
  geneList,
  simplify = T,
  input = "SYMBOL",
  method = c("ORA", "GSEA"),
  background = c("GO", "KEGG", "MKEGG", "Reactome")
)
}
\arguments{
\item{gene}{SYMBOL vector for ORA}

\item{geneList}{Named numeric vector for GSEA (names match input)}

\item{simplify}{Whether to simplify GO results}

\item{input}{"SYMBOL" or "ENTREZID" (for KEGG/MKEGG/Reactome)}

\item{method}{c("ORA","GSEA"), supports multiple}

\item{background}{c("GO","KEGG","MKEGG","Reactome"), supports multiple}
}
\value{
A list of enrichment results per method-background combo
}
\description{
Enrichment analysis for GO/KEGG/MKEGG/Reactome using ORA or GSEA; iterate method first, then background.
}
\examples{
\dontrun{
#' @examples
# We here use SYMBOL in the first place
library(org.Hs.eg.db)
data(geneList, package="DOSE")
gene.df <- tibble(logFC=geneList, ID=names(geneList))
gene.df.map <- bitr(gene.df$ID, fromType = "ENTREZID", # optional
                    toType = c("SYMBOL"),
                    OrgDb = org.Hs.eg.db)
gene.df <- gene.df \%>\% left_join(gene.df.map, by = c("ID" = "ENTREZID")) \%>\% drop_na()
gene.df <- gene.df \%>\% arrange(desc(logFC))
gene <- gene.df$SYMBOL[1:100] # for ORA
geneList <- setNames(gene.df$logFC, gene.df$SYMBOL)  # for GSEA
}
}
