% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.misc_fav.R
\name{leo_log}
\alias{leo_log}
\title{Log messages with time stamps}
\usage{
leo_log(..., level = "info", verbose = TRUE)
}
\arguments{
\item{...}{The message string to log, which will be pasted together.}

\item{level}{The log level. Options are \code{"info"}, \code{"success"}, \code{"warning"}, and \code{"danger"}.}

\item{verbose}{in case you want to turn the info log off, set it to FALSE.}
}
\value{
No return value. Outputs a formatted log message with a timestamp.
}
\description{
Logs messages with time stamps
The messages are styled using the \code{cli} package for enhanced readability.
This function can not deal with {} function in the \code{cli} package.
}
\examples{
n1 <- 10; n2 <- 20
leo_log("Processing the", n1, "and", n2, "files.")
leo_log("Task completed successfully!", level = "success")
leo_log("Potential issue detected.", level = "warning")
leo_log("Error occurred during processing!", level = "danger")
}
