% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.R
\name{leo_theme}
\alias{leo_theme}
\title{Custom ggplot2 Theme}
\usage{
leo_theme(
  plot_title_size = 14,
  legend_title_size = 10,
  legend_text_size = 9,
  axis_title_size = 10,
  axis_text_size = 8,
  type = c("console", "object")
)
}
\arguments{
\item{plot_title_size}{Numeric; title font size (default 14).}

\item{legend_title_size}{Numeric; legend title font size (default 10).}

\item{legend_text_size}{Numeric; legend text size (default 9).}

\item{axis_title_size}{Numeric; axis title size (default 10).}

\item{axis_text_size}{Numeric; axis text size (default 8).}

\item{type}{Character; "object" to return theme, or "console" (default) to print copyable code.}
}
\description{
A concise ggplot2 theme with nature-style sizing and black fonts.
}
\examples{
library(ggplot2)
# Add to a plot
ggplot(mtcars, aes(wt, mpg)) + geom_point() +
  leo_theme()
# Print theme code
leo_theme(type = "console")
}
