% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3.plot.R
\name{put_legend_inside}
\alias{put_legend_inside}
\title{Put legend inside ggplot/patchwork at (0.8, 0.8) npc}
\usage{
put_legend_inside(p, x = 0.8, y = 0.8)
}
\arguments{
\item{p}{ggplot object, list of ggplot objects, or patchwork}

\item{x, y}{npc coordinates in \link{0,1}, default 0.8, 0.8}
}
\value{
Same type as input with legend repositioned
}
\description{
Works for a single ggplot, a list of ggplots, or a patchwork object.
}
