% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3.plot.R
\name{rasterize_layers}
\alias{rasterize_layers}
\title{Rasterize point-like layers in a ggplot object}
\usage{
rasterize_layers(
  plot,
  dpi = 300,
  layers = c("Point", "Jitter", "Line", "Segment", "Linerange", "EdgeSegment")
)
}
\arguments{
\item{plot}{A ggplot object.}

\item{dpi}{Integer DPI for rasterization (default 300).}

\item{layers}{Character vector of layer types to rasterize, e.g. c("Point","Jitter","Line","Segment","EdgeSegment").}
}
\value{
ggplot object with target layers rasterized.
}
\description{
Rasterize point-like layers in a ggplot object
}
\examples{
library(ggplot2); library(ggrastr)
plot <- ggplot(diamonds, aes(carat, price, colour = cut)) + geom_point()
rasterize_layers(plot, dpi = 100)
}
