% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0.misc.R
\name{set_proxy}
\alias{set_proxy}
\title{Set or unset proxy}
\usage{
set_proxy(enable = TRUE, host = "127.0.0.1", port = 7897, socks = TRUE)
}
\arguments{
\item{enable}{TRUE to enable, FALSE to disable.}

\item{host}{Proxy host address.}

\item{port}{Proxy port number.}

\item{socks}{Whether to also set SOCKS5 proxy.}
}
\description{
Quickly enable or disable HTTP/HTTPS (and optionally SOCKS5) proxy for the current R session.
}
\examples{
# Enable proxy (HTTP/HTTPS + SOCKS5)
set_proxy(TRUE, host = "127.0.0.1", port = 7897, socks = TRUE)

# Test proxy
system("curl cip.cc")

# Disable proxy
set_proxy(FALSE)
}
