% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.misc_fav.R
\name{vd}
\alias{vd}
\title{View DataFrame with Visidata from Terminal}
\usage{
vd(df)
}
\arguments{
\item{df}{Dataframe or matrix to view.}
}
\description{
Opens a dataframe in Visidata (\code{vd}) from the R console.
Uses \code{data.table} or \code{vroom} for fast writing if available.
}
\examples{
\dontrun{
  data(mtcars)
  vd(mtcars)
  # Pipe support
  mtcars \%>\% vd()
}
}
