% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0.misc.R
\name{install_local}
\alias{install_local}
\title{Install local package}
\usage{
install_local(path)
}
\arguments{
\item{path}{The path to the local package source.}
}
\description{
This function is writen as I often forget what to code when I want to install a package from local source.
It is in fact quite simple to add "repos = NULL, type = "source" in \code{install.packages()}.
Anyway, it is always good to simplify the process.
}
